added login function to edit

+++ controllers/api/comment-routes.js
@@ -1,27 +1,22 @@
 const router = require('express').Router();
 const { Comment } = require('../../models');
 
-// GET api/comments
 router.get('/', (req, res) => {
-    Comment.findAll({
-        attributes: ['id', 'comment_text', 'user_id', 'post_id'],
-      })
-        .then(dbCommentData => res.json(dbCommentData))
-        .catch(err => {
-          console.log(err);
-          res.status(500).json(err);
-        });
-      });
-    
-// POST api/comments
+  Comment.findAll()
+    .then(dbCommentData => res.json(dbCommentData))
+    .catch(err => {
+      console.log(err);
+      res.status(500).json(err);
+    });
+});
+
 router.post('/', (req, res) => {
-  // check the session
+  // expects => {comment_text: "This is the comment", user_id: 1, post_id: 2}
   if (req.session) {
     Comment.create({
       comment_text: req.body.comment_text,
-      post_id: req.body.post_id,
-      // use the id from the session
-      user_id: req.session.user_id
+      user_id: req.session.user_id,
+      post_id: req.body.post_id
     })
       .then(dbCommentData => res.json(dbCommentData))
       .catch(err => {
@@ -31,24 +26,25 @@ router.post('/', (req, res) => {
   }
 });
 
-//DELETE api/comments/:id
 router.delete('/:id', (req, res) => {
+  if (req.session) {
     Comment.destroy({
-        where: {
-          id: req.params.id
+      where: {
+        id: req.params.id
+      }
+    })
+      .then(dbCommentData => {
+        if (!dbCommentData) {
+          res.status(404).json({ message: 'No comment found with this id!' });
+          return;
         }
+        res.json(dbCommentData);
       })
-        .then(dbCommentData => {
-          if (!dbCommentData) {
-            res.status(404).json({ message: 'No comment found with this id' });
-            return;
-          }
-          res.json(dbCommentData);
-        })
-        .catch(err => {
-          console.log(err);
-          res.status(500).json(err);
-        });
-    });
+      .catch(err => {
+        console.log(err);
+        res.status(500).json(err);
+      });
+  }
+});
 
 module.exports = router;
